Refactor epoch handling and PIT import logic
Standardizes epoch fields to int64 throughout the codebase, updates PIT import routines for robust multi-antenna mapping, and improves deterministic filtering in time_to_event. Adds orphan tag detection to radio_project and scripts, and introduces new diagnostic scripts for TTE and tag inspection. Updates project paths and receiver types in scripts for new dataset.

+++ pymast/overlap_removal.py
@@ -221,7 +221,7 @@ class bout():
         rec_dat = rec_dat[['freq_code', 'epoch', 'time_stamp', 'power', 'rec_id']]
         rec_dat = rec_dat.astype({
             'freq_code': 'object',
-            'epoch': 'float32',
+            'epoch': 'int64',
             'time_stamp': 'datetime64[ns]',
             'power': 'float32',
             'rec_id': 'object'
@@ -312,7 +312,7 @@ class bout():
         presence_df = self.presence_df.astype({
             'freq_code': 'object',
             'rec_id': 'object',
-            'epoch': 'float32',
+            'epoch': 'int64',
             'time_stamp': 'datetime64[ns]',
             'power': 'float32',
             'bout_no': 'int32',
